white_check_list: Fix some forging test
